v2 reporter: changes after pr review
